requirements.txt
hydra-core==0.11.3
torch==1.3.1
torchsummary==1.5.1
torchvision==0.4.2
tqdm==4.42.1
